Task 0

+++ 0x15-file_io/0-read_textfile.c
@@ -42,7 +42,7 @@ ssize_t read_textfile(const char *filename, size_t letters)
 	}
 
 	n_written = write(STDOUT_FILENO, buffer, n_read);
-	if (n_written == -1 || (size_t)n_written != n_read)
+	if (n_written == -1 || (ssize_t)n_written != (ssize_t)n_read)
 	{
 		free(buffer);
 		close(fd);
